refactor: relay and system staking page
